requirements.txt
cv2==4.3.0
flask==1.1.2
keras==2.4.3
matplotlib==3.2.1
numpy==1.18.1
Pillow==7.2.0
re==1.4.4
seaborn==0.10.0
sklearn==0.23.1
tensorflow==2.2.0
WTForms==2.3.3
Flask-WTF==0.14.3